Add centralised debug logging
- Introduces a shared debugging utility that reads the debug flag from the `SURREAL_DEBUG` environment variable and exposes a helper for conditional logging
- Replaces file-level `DEBUG` constants in the SurrealDB node with an import from the new shared module
- Updates helper utilities and operations to use the centralized debug flag and remove redundant per-file declarations
