+++ src/nodes/SurrealDb/resources/index/operations/createIndex.operation.ts
@@ -14,8 +14,7 @@ import {
 import { debugLog, addSuccessResult } from "../../../utilities";
 import type { IOperationHandler } from "../../../types/operation.types";
 
-// Set to true to enable debug logging, false to disable
-const DEBUG = false;
+import { DEBUG } from '../../../debug';
 
 /**
  * Implementation of the "Create Index" operation
