+++ src/nodes/SurrealDb/resources/record/operations/createRecord.operation.ts
@@ -18,8 +18,7 @@ import {
   addSuccessResult,
 } from "../../../utilities";
 
-// Set to true to enable debug logging, false to disable
-const DEBUG = false;
+import { DEBUG } from '../../../debug';
 
 /**
  * Create Record operation handler for Record resource
