+++ src/nodes/SurrealDb/resources/record/operations/deleteRecord.operation.ts
@@ -14,8 +14,7 @@ import {
   addErrorResult,
 } from "../../../utilities";
 
-// Set to true to enable debug logging, false to disable
-const DEBUG = false;
+import { DEBUG } from '../../../debug';
 
 /**
  * Delete Record operation handler for Record resource
