+++ src/nodes/SurrealDb/resources/record/operations/getRecord.operation.ts
@@ -13,8 +13,7 @@ import {
   debugLog,
 } from "../../../utilities";
 
-// Set to true to enable debug logging, false to disable
-const DEBUG = false;
+import { DEBUG } from '../../../debug';
 
 /**
  * Get Record operation handler for Record resource
