+++ src/nodes/SurrealDb/resources/record/operations/mergeRecord.operation.ts
@@ -18,8 +18,7 @@ import {
   addErrorResult,
 } from "../../../utilities";
 
-// Set to true to enable debug logging, false to disable
-const DEBUG = false;
+import { DEBUG } from '../../../debug';
 
 /**
  * Merge Record operation handler for Record resource
