+++ src/nodes/SurrealDb/resources/record/operations/updateRecord.operation.ts
@@ -19,8 +19,7 @@ import {
   addErrorResult,
 } from "../../../utilities";
 
-// Set to true to enable debug logging, false to disable
-const DEBUG = false;
+import { DEBUG } from '../../../debug';
 
 /**
  * Update Record operation handler for Record resource
