+++ src/nodes/SurrealDb/resources/record/operations/upsertRecord.operation.ts
@@ -14,8 +14,7 @@ import {
   addSuccessResult,
 } from "../../../utilities";
 
-// Set to true to enable debug logging, false to disable
-const DEBUG = false;
+import { DEBUG } from '../../../debug';
 
 /**
  * Upsert Record operation handler for Record resource
