+++ src/nodes/SurrealDb/resources/table/operations/createMany.operation.ts
@@ -8,8 +8,7 @@ import {
 } from "../../../GenericFunctions";
 import type { IOperationHandler } from "../../../types/operation.types";
 
-// Set to true to enable debug logging, false to disable
-const DEBUG = false;
+import { DEBUG } from '../../../debug';
 
 /**
  * Implementation of the "Create Many Records" operation
