+++ src/nodes/SurrealDb/resources/table/operations/deleteAllRecords.operation.ts
@@ -18,8 +18,7 @@ import {
 } from "../../../GenericFunctions";
 import type { IOperationHandler } from "../../../types/operation.types";
 
-// Set to true to enable debug logging, false to disable
-const DEBUG = false;
+import { DEBUG } from '../../../debug';
 
 /**
  * Implementation of the "Delete All Records" operation
