+++ src/nodes/SurrealDb/resources/table/operations/getAllRecords.operation.ts
@@ -18,8 +18,7 @@ import {
 } from "../../../GenericFunctions";
 import type { IOperationHandler } from "../../../types/operation.types";
 
-// Set to true to enable debug logging, false to disable
-const DEBUG = false;
+import { DEBUG } from '../../../debug';
 
 /**
  * Implementation of the "Get All Records" operation
