+++ src/nodes/SurrealDb/resources/table/operations/getMany.operation.ts
@@ -19,8 +19,7 @@ import {
 } from "../../../GenericFunctions";
 import type { IOperationHandler } from "../../../types/operation.types";
 
-// Set to true to enable debug logging, false to disable
-const DEBUG = false;
+import { DEBUG } from '../../../debug';
 
 /**
  * Implementation of the "Get Many Records" operation
