+++ src/nodes/SurrealDb/resources/table/operations/listTables.operation.ts
@@ -11,8 +11,7 @@ import {
 import { debugLog, addErrorResult } from "../../../utilities";
 import type { IOperationHandler } from "../../../types/operation.types";
 
-// Set to true to enable debug logging, false to disable
-const DEBUG = false;
+import { DEBUG } from '../../../debug';
 
 /**
  * Implementation of the "List Tables" operation
